Split compiler sources into three: parser, tests, compiler
